+++ TourBackend/TourBackend/TourBackend/Protocols/ControlToSyncActorProtocol.cs
@@ -39,10 +39,10 @@ namespace TourBackend
         public bool isActive = true;
         public string objectid;
         public int mediaid;
-        public int[] position;
-        public int[] rotation;
+        public float[] position;
+        public float[] rotation;
 
-        public CodeObject(string _objectid, int _mediaid, int[] _position, int[] _rotation, bool _isActive) {
+        public CodeObject(string _objectid, int _mediaid, float[] _position, float[] _rotation, bool _isActive) {
             objectid = _objectid;
             mediaid = _mediaid;
             position = _position;
@@ -50,7 +50,7 @@ namespace TourBackend
             isActive = _isActive;
         }
 
-        public CodeObject(string _objectid, int _mediaid, int[] _position, int[] _rotation)
+        public CodeObject(string _objectid, int _mediaid, float[] _position, float[] _rotation)
         {
             objectid = _objectid;
             mediaid = _mediaid;
