+++ TourBackend/TourBackend/TourBackend/Protocols/RecognitionManagerToCodeObjectsProtocol.cs
@@ -18,10 +18,10 @@ namespace TourBackend
         public string messageid;
         public string objectid;
         public int mediaid;
-        public int[] position;
-        public int[] rotation;
+        public float[] position;
+        public float[] rotation;
 
-        public UpdateCodeObjectActor(string _messageid, string _objectid, int _mediaid, int[] _position, int[] _rotation) {
+        public UpdateCodeObjectActor(string _messageid, string _objectid, int _mediaid, float[] _position, float[] _rotation) {
             messageid = _messageid;
             mediaid = _mediaid;
             position = _position;
